// src/FrontEnd/src/app/models/employee/CreateEmployeeCommand.model.ts
export class CreateEmployeeCommand {
    FirstName: string;
    LastName: string;
    TypePerson: number;
    CountryRegistryNumber: string;
    StateRegistryNumber: string;
    CityRegistryNumber: string;
    PhoneNumber1: string;
    PhoneNumber2: string;
    MobilePhoneNumber1: string;
    MobilePhoneNumber2: string;
    EmailAddress: string;
    EmailAdminSystem: string;
    AditionalInfo: string;
    UserName: string;
    Password: string;
    ConfirmPassword: string;
    UserProfile: string;
    Subsidiary: string;
    Position: string;
    CostCenterArea: string;
    Company: string;
}